Reorganize transaction type traits
