+++ crates/cosmos/cosmos-client-components/src/impls/transaction/query_nonce.rs
@@ -1,6 +1,8 @@
 use cgp_core::CanRaiseError;
 use hermes_relayer_components::transaction::traits::nonce::query_nonce::NonceQuerier;
-use hermes_relayer_components::transaction::traits::types::{HasNonceType, HasSignerType};
+use hermes_relayer_components::transaction::traits::types::{
+    nonce::HasNonceType, signer::HasSignerType,
+};
 use ibc_relayer::chain::cosmos::query::account::query_account;
 use ibc_relayer::chain::cosmos::types::account::Account;
 use ibc_relayer::error::Error as RelayerError;
