+++ crates/cosmos/cosmos-client-components/src/impls/transaction/query_tx_response.rs
@@ -1,7 +1,7 @@
 use cgp_core::CanRaiseError;
 use hermes_relayer_components::transaction::traits::components::tx_response_querier::TxResponseQuerier;
 use hermes_relayer_components::transaction::traits::types::{
-    HasTransactionHashType, HasTxResponseType,
+    tx_hash::HasTransactionHashType, tx_response::HasTxResponseType,
 };
 use ibc_relayer::chain::cosmos::query::tx::query_tx_response;
 use ibc_relayer::error::Error as RelayerError;
