+++ crates/cosmos/cosmos-client-components/src/impls/transaction/submit_tx.rs
@@ -3,7 +3,7 @@ use core::fmt::Debug;
 use cgp_core::CanRaiseError;
 use hermes_relayer_components::transaction::traits::components::tx_submitter::TxSubmitter;
 use hermes_relayer_components::transaction::traits::types::{
-    HasTransactionHashType, HasTransactionType,
+    transaction::HasTransactionType, tx_hash::HasTransactionHashType,
 };
 use ibc_proto::cosmos::tx::v1beta1::TxRaw;
 use ibc_relayer::chain::cosmos::types::tx::SignedTx;
