+++ crates/relayer/relayer-components/src/transaction/traits/components/send_messages_with_signer.rs
@@ -4,7 +4,7 @@ use cgp_core::prelude::*;
 
 use crate::chain::traits::types::event::HasEventType;
 use crate::chain::traits::types::message::HasMessageType;
-use crate::transaction::traits::types::HasSignerType;
+use crate::transaction::traits::types::signer::HasSignerType;
 
 #[derive_component(MessagesWithSignerSenderComponent, MessagesWithSignerSender<Chain>)]
 #[async_trait]
