+++ crates/relayer/relayer-components/src/transaction/traits/components/send_messages_with_signer_and_nonce.rs
@@ -1,7 +1,8 @@
 use cgp_core::prelude::*;
 
 use crate::chain::traits::types::message::HasMessageType;
-use crate::transaction::traits::types::{HasNonceType, HasSignerType, HasTxResponseType};
+use crate::transaction::traits::types::nonce::HasNonceType;
+use crate::transaction::traits::types::{signer::HasSignerType, tx_response::HasTxResponseType};
 
 #[derive_component(MessagesWithSignerAndNonceSenderComponent, MessagesWithSignerAndNonceSender<TxContext>)]
 #[async_trait]
