+++ crates/relayer/relayer-components/src/transaction/traits/components/tx_submitter.rs
@@ -1,6 +1,8 @@
 use cgp_core::prelude::*;
 
-use crate::transaction::traits::types::{HasTransactionHashType, HasTransactionType};
+use crate::transaction::traits::types::{
+    transaction::HasTransactionType, tx_hash::HasTransactionHashType,
+};
 
 #[derive_component(TxSubmitterComponent, TxSubmitter<TxContext>)]
 #[async_trait]
