+++ crates/relayer/relayer-components/src/transaction/traits/event.rs
@@ -3,7 +3,7 @@ use alloc::vec::Vec;
 use cgp_core::prelude::*;
 
 use crate::chain::traits::types::event::HasEventType;
-use crate::transaction::traits::types::HasTxResponseType;
+use crate::transaction::traits::types::tx_response::HasTxResponseType;
 
 #[derive_component(TxResponseAsEventsParserComponent, TxResponseAsEventsParser<Chain>)]
 pub trait CanParseTxResponseAsEvents: HasTxResponseType + HasEventType + HasErrorType {
