+++ crates/relayer/relayer-components/src/transaction/traits/logs/nonce.rs
@@ -1,6 +1,6 @@
 use crate::logger::traits::has_logger::HasLoggerType;
 use crate::logger::traits::logger::BaseLogger;
-use crate::transaction::traits::types::HasNonceType;
+use crate::transaction::traits::types::nonce::HasNonceType;
 
 pub trait CanLogNonce: HasNonceType + HasLoggerType {
     fn log_nonce<'a>(nonce: &'a Self::Nonce) -> <Self::Logger as BaseLogger>::LogValue<'a>;
