+++ crates/relayer/relayer-components/src/transaction/traits/nonce/allocate_nonce.rs
@@ -1,7 +1,7 @@
 use cgp_core::prelude::*;
 
 use crate::transaction::traits::nonce::guard::HasNonceGuard;
-use crate::transaction::traits::types::HasSignerType;
+use crate::transaction::traits::types::signer::HasSignerType;
 
 #[derive_component(NonceAllocatorComponent, NonceAllocator<TxContext>)]
 #[async_trait]
