+++ crates/relayer/relayer-components/src/transaction/traits/nonce/mutex.rs
@@ -3,7 +3,7 @@ use cgp_core::prelude::*;
 use crate::runtime::traits::mutex::{HasMutex, MutexGuardOf, MutexOf};
 use crate::runtime::traits::runtime::HasRuntime;
 use crate::transaction::traits::nonce::guard::HasNonceGuard;
-use crate::transaction::traits::types::HasSignerType;
+use crate::transaction::traits::types::signer::HasSignerType;
 
 /**
    A naive nonce allocator that simply query the current nonce from the context
