+++ crates/relayer/relayer-components/src/transaction/traits/nonce/query_nonce.rs
@@ -1,6 +1,7 @@
 use cgp_core::prelude::*;
 
-use crate::transaction::traits::types::{HasNonceType, HasSignerType};
+use crate::transaction::traits::types::nonce::HasNonceType;
+use crate::transaction::traits::types::signer::HasSignerType;
 
 #[derive_component(NonceQuerierComponent, NonceQuerier<TxContext>)]
 #[async_trait]
