+++ crates/relayer/relayer-components/src/transaction/traits/types/fee.rs
@@ -1,0 +1,6 @@
+use cgp_core::prelude::*;
+
+#[derive_component(FeeTypeComponent, ProvideFeeType<Chain>)]
+pub trait HasFeeType: Async {
+    type Fee: Async;
+}
