+++ crates/relayer/relayer-components/src/transaction/traits/types/nonce.rs
@@ -1,0 +1,6 @@
+use cgp_core::prelude::*;
+
+#[derive_component(NonceTypeComponent, ProvideNonceType<Chain>)]
+pub trait HasNonceType: Async {
+    type Nonce: Async;
+}
