+++ crates/relayer/relayer-components/src/transaction/traits/types/signer.rs
@@ -1,0 +1,8 @@
+use cgp_core::prelude::*;
+
+#[derive_component(SignerTypeComponent, ProvideSignerType<Chain>)]
+pub trait HasSignerType: Async {
+    type Signer: Async;
+}
+
+pub type SignerOf<Context> = <Context as HasSignerType>::Signer;
