+++ crates/relayer/relayer-components/src/transaction/traits/types/transaction.rs
@@ -1,0 +1,8 @@
+use cgp_core::prelude::*;
+
+#[derive_component(TransactionTypeComponent, ProvideTransactionType<Chain>)]
+pub trait HasTransactionType: Async {
+    type Transaction: Async;
+
+    fn tx_size(tx: &Self::Transaction) -> usize;
+}
