+++ crates/relayer/relayer-components/src/transaction/traits/types/tx_hash.rs
@@ -1,0 +1,6 @@
+use cgp_core::prelude::*;
+
+#[derive_component(TransactionHashTypeComponent, ProvideTransactionHashType<Chain>)]
+pub trait HasTransactionHashType: Async {
+    type TxHash: Async;
+}
