+++ crates/relayer/relayer-components/src/transaction/traits/types/tx_response.rs
@@ -1,0 +1,6 @@
+use cgp_core::prelude::*;
+
+#[derive_component(TxResponseTypeComponent, ProvideTxResponseType<Chain>)]
+pub trait HasTxResponseType: Async {
+    type TxResponse: Async;
+}
