+++ crates/sovereign/sovereign-client-components/src/sovereign/impls/transaction/query_tx_response.rs
@@ -3,7 +3,7 @@ use core::ops::Range;
 use cgp_core::CanRaiseError;
 use hermes_relayer_components::transaction::traits::components::tx_response_querier::TxResponseQuerier;
 use hermes_relayer_components::transaction::traits::types::{
-    HasTransactionHashType, HasTxResponseType,
+    tx_hash::HasTransactionHashType, tx_response::HasTxResponseType,
 };
 use hex::ToHex;
 use jsonrpsee::core::client::ClientT;
