+++ crates/sovereign/sovereign-cosmos-relayer/src/contexts/sovereign_rollup.rs
@@ -25,12 +25,12 @@ use hermes_relayer_components::transaction::traits::components::tx_response_poll
 use hermes_relayer_components::transaction::traits::components::tx_response_querier::CanQueryTxResponse;
 use hermes_relayer_components::transaction::traits::components::tx_response_querier::TxResponseQuerierComponent;
 use hermes_relayer_components::transaction::traits::event::TxResponseAsEventsParserComponent;
-use hermes_relayer_components::transaction::traits::types::FeeTypeComponent;
-use hermes_relayer_components::transaction::traits::types::NonceTypeComponent;
-use hermes_relayer_components::transaction::traits::types::SignerTypeComponent;
-use hermes_relayer_components::transaction::traits::types::TransactionHashTypeComponent;
-use hermes_relayer_components::transaction::traits::types::TransactionTypeComponent;
-use hermes_relayer_components::transaction::traits::types::TxResponseTypeComponent;
+use hermes_relayer_components::transaction::traits::types::fee::FeeTypeComponent;
+use hermes_relayer_components::transaction::traits::types::nonce::NonceTypeComponent;
+use hermes_relayer_components::transaction::traits::types::signer::SignerTypeComponent;
+use hermes_relayer_components::transaction::traits::types::transaction::TransactionTypeComponent;
+use hermes_relayer_components::transaction::traits::types::tx_hash::TransactionHashTypeComponent;
+use hermes_relayer_components::transaction::traits::types::tx_response::TxResponseTypeComponent;
 use hermes_relayer_runtime::impls::logger::components::ProvideTracingLogger;
 use hermes_relayer_runtime::impls::types::runtime::ProvideTokioRuntimeType;
 use hermes_relayer_runtime::types::runtime::HermesRuntime;
